+++ custom_components/blueiris/helpers/const.py
@@ -176,3 +176,9 @@ CONF_CLEAR_CREDENTIALS = "clear-credentials"
 
 DOMAIN_LOAD = "load"
 DOMAIN_UNLOAD = "unload"
+
+CONF_CONTENT_TYPE = "content_type"
+CONF_LIMIT_REFETCH_TO_URL_CHANGE = "limit_refetch_to_url_change"
+CONF_STILL_IMAGE_URL = "still_image_url"
+CONF_STREAM_SOURCE = "stream_source"
+CONF_FRAMERATE = "framerate"